+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Fix false positive in `Style/TrailingCommaInArguments` & `Style/TrailingCommaInLiteral` cops with consistent_comma style. ([@meganemura][])
 * [#2861](https://github.com/bbatsov/rubocop/pull/2861): Fix false positive in `Style/SpaceAroundKeyword` for `rescue(...`. ([@rrosenblum][])
 * [#2832](https://github.com/bbatsov/rubocop/issues/2832): `Style/MultilineOperationIndentation` treats operations inside blocks inside other operations correctly. ([@jonas054][])
+* [#2865](https://github.com/bbatsov/rubocop/issues/2865): Change `require:` in config to be relative to the `.rubocop.yml` file itself. ([@ptarjan][])
 
 ## 0.37.2 (11/02/2016)
 
@@ -2002,3 +2003,4 @@
 [@mmcguinn]: https://github.com/mmcguinn
 [@pocke]: https://github.com/pocke
 [@prsimp]: https://github.com/prsimp
+[@ptarjan]: https://github.com/ptarjan